Add module to manage reading from screen
Also set up a test to verify that it's able to capture the current
client's window.
Having some issues with dual monitors, especially because my second
monitor is a different resolution to the first. I can screen grab from
it, but the image isn't the correct size. As long as I screengrab from
the primary monitor it works fine through.