Fix screen grabbing from different monitors
With this fix we can screen grab the client even if it has a different
resolution to the primary monitor.

Still a problem with scaling the config values, because they were set up
for my primary monitor, and for some reason don't work correctly on the
second monitor - However grabbing the whole window works, so there must
be some scaling required.