Add scaling factor todo
Finally figured out why the screen grabs were coming 50% larger on
the first monitor... it's also because of DPI scaling.

I think I should be able to collect the scale factor and apply that
to config values, but it'll be a bit annoying to use config values
that way, so for the time being I'm just assuming scale factor is
100% (I've set it at the system level) and then using the config values
directly.