Add methods to identify items in each inventory slot
Inventory items are matched against templates stored in module data.
If templates haven't been saved out it won't be able to identify
anything and will throw a warning.

I also added a default if no template names have been specified on load
to attempt to find numpy files in the standard data location. This
allows templates to be configured, but also has a default for
convenience.

Currently running a whole inventory match in ~0.05 seconds, which is
pretty good. Next step is to have that working asynchronously in the
background. This will involve setting up a main update loop, where the
client is screen grabbed once per loop, and then sections of the
screengrab can be passed to various functions (including inventory
slots).