Add template matching to game objects base class
All GameObjects can now be identified with a standard template matching
method, as long as templates are specified on init.

Several existing subclasses of GameObject needed to be refactored in
order to work properly with this. Notably, the inventory and spells got
a bit messed up because they have custom interfaces that I needed to
create a mixin to work.

Logout menu button now has those templates defined.