Add logout method implementation in bone offering script
This works as long as the current logout menu is the normal logout menu,
and not the world switcher context.
I'll need to account for both, as world switching to pvp worlds is
useful for banking, but being in a non-pvp world means you're less
likely to run into someone.