Add feature matching experiments for feature matching minimap
Trying various methods to run feature matching on the minimap against
the map sections. Seems to work reasonably well on a single map section,
but doesn't work so well on the larger map (which is required because
there's a chance we'll need to go slightly off a single section.
It also didn't work at all well in the chaos alter section, probably
because there's not enough distinct features on that map set for it work
with.