Attempt to filter feature matching results
This didn't work out so well, but the idea was to try to find the
section of the map with the most concurrent features, weighted by their
average, normalised distance, so that we would tend towards the section
that contains more of the correct features.

It didn't really work as expected, often just ending up with a square
containing only one good match (meaning the average is, by default,
lower) and even then, that one match is often in completely the wrong
place.