Add edge case capture for normalise method
Since we're going to pass in keypoint values, i.e. programatically
determined values instead of known values, in theory the min and max
values could be anything, so it's possible that they're the same, which
would cause a ZeroDivisionError and make the program crash.

Capturing this exception means we can move on with our lives.