Add various match filtering methods
The issue I was having was mostly because so many of the features picked
up by the algorithm are completely wrong, so they make it difficult to
figure out anything simple.

Added various filtering methods to try to work through to a useable
solution;
 - Manual Filtering: As the name suggests, manual filtering just means I
   can go through by hand and pick only the correctly identified
   features. This was useful because it allowed me to carefully see what
   percentage of matches actually worked.
 - By Distance: This is a very simple filter picking only the values
   that the orb algorithm determines to be 'best'. In reality, this
   didn't work at all.
 - Top X Results: Similar to by distance, this just sorts by distance
   and then choose the top x. This sometimes worked, but often even the
   top match was way off.
 - Grouped Matches: This was by far the most successful method. The idea
   here was that each feature implies a position the minimap should be
   in relation to the main map. If a group of matches are all correct,
   that implied position should be the same. The number of matches in a
   group is normalised and divides the average distance, which should
   give the best match being where the distance is lower, and the number
   of matches is higher.

In all cases, the algorithms would more correctly identify the position
when there was less main map to compare against. Nameely, just using the
ghorrock teleport map chunk only actually worked, but as soon as the
rest of the map is included, it becomes wildly inaccurate.

I think the next step would be to create some method of trimming the
section of the main map around the current position, but with a bit of a
margin to allow movement.