Add feature matching methods to minimap class
Minimap now has methods that allow us to sample the current minimap,
perform feature matching on it against standard loaded maps in order to
determine the player's current location within the map.

Also added logging to game objects with a simple console logger. Feature
matching against the runescape maps was pretty complex, so logging was
required to make sense of what was happening.

Something is wrong with the local zone calculation, however, because the
local zone seems to slide even if the position is drawn correctly on the
shown map. This eventually spins out of control and off the map
entirely, before throwing an error, so ideally we'd like to get this
fixed :)