Implement CRST tracker for player entity
Seems to work OK, but gets easily confused if an NPC (or other player
probably) passes over the tracking bounding box. Once it starts tracking
the other object it doesn't seem to self correct either... There's
probably a way to mitigate that, but that will have to be for another
time.