Enable edgeless nodes
Can be used for things key locations on the map (e.g. bank booths) that
are not actually accessible to stand on, but still need to be found.

Also updated find nearest method in Map, because with edgeless nodes, if
we're standing next to a bank booth, that bank booth will be the nearest
node ... but there is no path from it to any other location, so pathing
would break.