Improve click condition for target fire
Because we're determining fires being lit by xp drops, at the exact
point the player is still on the original tile, so there was always a
click on the next tile, even though fire making auto-pushes the player
to the next tile, so a click is not necessary.

Depending on the timing this was sometimes causing the nest west tile to
be clicked, then te fire lit of that tile, but registered as on the east
tile. So when the player made it to the west tile it would try to light
a log, but couldn't because there's already fire there.

Conditions are now expanded to include if xp has been dropped that we
should be waiting to arrive, instead of clicking straight away (which we
can't because we're in an animation stall).

Also adjusted how the key is updated on the fires, because I was finding
very often the key would be out of step with the map. Now, if the gps
results are not accepted (per the tree positions) then the fire
positions are not updated either. This is working so far, but will break
down once we start banking because the trees will be out of range.