Reset targets on state change
While firemaking we shouldn't have a target tree, because we won't be
clicking them again until we're firemaking.

Conversely, while woodcutting, we don't need target fires.

And then while banking, we don't need any target fires or trees.
(We'll need a target booth when we get to it though...)

Related to this, tree colour needs to be set back to normal if not the
target.