Add method to recalculate static entity screen position
Instead of pattern matching and filtering coordinates like I was in the
lumberjack script, this method is to assume the map coordinates are
correct and re-calculate to the screen position relative to them.

It also assume the entity's global coordinates haven't changed, so isn't
really appropriate for NPCs (although could be a useful step in tracking
their movement).