Add histogram cross-referencing to minimap
Minimap templates are so small that using masks causes many false
positives (I don't know if this is a bug in cv2 or my use of it, but
it's a problem either way...).

Histograms are another method to compare the candidate image swatches
against the original template, enabling me to filter out the false
positives and still use masks.

Using a mask was especially important when an item/NPC is on the
northern border of a tree, and the item/NPC minimap icon actually spills
over into the tile south of it. Using a mask on the trees allows us to
identify them even if partially occluded.