Allow game object to click without moving
Was finding that clicking game objects would always force a move within
the bounding box, even if the current mouse position is already being
confirmed as clicking on the correct thing by checking the mouse options
(moving to the position of which is already normally distributed).

Sometimes this was causing the mouse to jump to an invalid position
without a chance to check the mouse options.

By passing in False to the click method, the mouse no longer moves
before clicking.