Bugfix to click entity return
Previously was returning mouse options on clicking the entity the first
time if there's no delay. But since we've just moved the mouse there's
no chance the mouse options would match, so it would always return
false.

With recent changes to clicking object outside of client, this is the
only fail condition we can check for now.