Convert gps to floating point precision
This greatly improves the ability to locate entities on the game screen,
and along with the floating point precision gps coordinates, we are now
also able to more accurately calculate the offset from the player.

In addition, entities are no longer calculated off the true tile, since
this was never really the case. The camera has a fixed 'sliding zone' in
the midde of the screen. Where the camera slides depends on where the
true tile is (e.g. true tile is NW of the fixed area, then the camera
will be slid NW, but no further).

There is also a minor drag offset applied, all of which meaning as long
as we get an accurate location on the map, we can now get a pretty
accurate location on entities on screen at all times, not just while
stationary.

Finally, in converting to this new system, all the existing maps have
some weird offsets, which is being stored in the map metadata.