Add target tracking
While attempting to click an object with the application methods,
objects will track their relative position and attempt to keep the mouse
at that same position.

This should increase the likelihood that the mouse stays inside the
object click box and we can make a successful click while on the move,
rather than what was previously happening where the mouse moves inside
initially, but still needs to wait for mouse options to confirm a click
is possible - but by the time the cycle comes round the mouse is already
outside the click box because the camera has moved.