Bug fix on game object click box
Click box was incorrectly maing the box inside the game screen if the
box was actually too far to one side. The intention here was just to
make sure the click box random coordinate selection was insde the client
bbox, but this was interfering with partial is_inside, so I've made it
so the adjusted bbox is 1 pixel outside the client, so random selection
should still generally work, but we don't get false positives.