Change sleep key to p
Was finding the capslock check while running the script was frequently
hitting race conditions and getting stuck in an on position.

Now the p key toggles sleep, but the capslock key is used to verify that
the toggling should occur (i.e. if we're sleeping, but we then need to
type the letter p).