Add consumable delay across consumables
Previously could get a click on a consumable, but it was not actually
consumed, because of a tick delay. This was especially a problem with
things like anitpoison, because it could mean a full 5-6 minutes
taking poison damage.

Also fixed a minor bug where recalculation was running on the old
assumption that object state would be None if the object has
disappeared, instead of 'nothing' or 'something' as is the case now.
Also since we're checking across all inv icons now, this
check is redundant anyway.